Adds HeuristicPlacePolicy class
